+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Release date: ?
  - refactored exceptions in Di
  - added addToContainerAfterBuild() to Backpack\AppBuilder
  - added multiple set type methods in Collection\Structure\DataType
+ - setProps() of Backpack\AppBuilder now accept an array or an instance of Blueprint\Collection\Dictionary
 
 VERSION 4.0-BETA1
 -----------